+++ framework/core/src/main/java/com/tenio/core/schedule/task/internal/CcuReportTask.java
@@ -1,7 +1,7 @@
 /*
 The MIT License
 
-Copyright (c) 2016-2022 kong <congcoi123@gmail.com>
+Copyright (c) 2016-2023 kong <congcoi123@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -58,16 +58,10 @@ public final class CcuReportTask extends AbstractSystemTask {
 
   @Override
   public ScheduledFuture<?> run() {
-    var threadFactory =
+    var threadFactoryTask =
         new ThreadFactoryBuilder().setDaemon(true).setNameFormat("ccu-report-task-%d").build();
-    return Executors.newSingleThreadScheduledExecutor(threadFactory).scheduleAtFixedRate(
-        () -> {
-          var worker = new Thread(() -> eventManager.emit(ServerEvent.FETCHED_CCU_INFO,
-              playerManager.getPlayerCount()));
-          worker.setDaemon(true);
-          worker.setName("ccu-report-worker");
-          worker.start();
-        },
+    return Executors.newSingleThreadScheduledExecutor(threadFactoryTask).scheduleAtFixedRate(
+        () -> eventManager.emit(ServerEvent.FETCHED_CCU_INFO, playerManager.getPlayerCount()),
         initialDelay, interval, TimeUnit.SECONDS);
   }