+++ framework/core/src/main/java/com/tenio/core/schedule/task/internal/SystemMonitoringTask.java
@@ -1,7 +1,7 @@
 /*
 The MIT License
 
-Copyright (c) 2016-2022 kong <congcoi123@gmail.com>
+Copyright (c) 2016-2023 kong <congcoi123@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -60,20 +60,14 @@ public final class SystemMonitoringTask extends AbstractSystemTask {
 
   @Override
   public ScheduledFuture<?> run() {
-    var threadFactory =
+    var threadFactoryTask =
         new ThreadFactoryBuilder().setDaemon(true).setNameFormat("system-monitoring-task-%d")
             .build();
-    return Executors.newSingleThreadScheduledExecutor(threadFactory).scheduleAtFixedRate(
-        () -> {
-          var worker = new Thread(
-              () -> eventManager.emit(ServerEvent.SYSTEM_MONITORING, systemMonitoring.getCpuUsage(),
-                  systemMonitoring.getTotalMemory(), systemMonitoring.getUsedMemory(),
-                  systemMonitoring.getFreeMemory(),
-                  systemMonitoring.countRunningThreads()));
-          worker.setDaemon(true);
-          worker.setName("system-monitoring-worker");
-          worker.start();
-        },
+    return Executors.newSingleThreadScheduledExecutor(threadFactoryTask).scheduleAtFixedRate(
+        () -> eventManager.emit(ServerEvent.SYSTEM_MONITORING, systemMonitoring.getCpuUsage(),
+            systemMonitoring.getTotalMemory(), systemMonitoring.getUsedMemory(),
+            systemMonitoring.getFreeMemory(),
+            systemMonitoring.countRunningThreads()),
         initialDelay, interval, TimeUnit.SECONDS);
   }
 }